requirements.txt
langchain
langchain-openai
openai
streamlit
streamlit-chat
python-dotenv
ipykernel